requirements.txt
django==3.0.7
djangorestframework==3.11.0
django-cors-headers==3.4.0
psycopg2-binary==2.8.4
celery==4.4.2
